refactor(website): unify EU service tabs into one page with query params
Collapse the per-section /pricing, /security, and /vs-{slug} routes
under each EU service into the canonical service URL with a `?tab=X`
query param. Tabs are now a shadcn Tabs primitive (proper WAI-ARIA
tab pattern) wrapping a sticky segmented control on the page.

Old paths are 308-redirected (single hop where possible, locale and
unprefixed forms covered) so existing links and indexed URLs keep
working. Sitemap, IndexNow URL collector, and OG metadata helper drop
the per-tab entries; LLM markdown gains the pricing-tier and
gdpr-notes sections that used to live on the sub-pages so the .md
endpoint still has the full picture.
